requirements.txt
pydantic
dotenv
pypandoc
google-genai
fastapi